front/requirements.txt
python-dateutil==2.8.2
python-dotenv==0.19.2
Flask==2.0.2
Flask-Caching==1.10.1
flask-login==0.5.0
Flask-WTF==1.0.0
Flask-Babel==2.0.0
wtforms[email]
requests==2.26.0
email-validator
jsonpickle
# development only
freezegun
pytest
